First pass at breakup
