+++ main/src/test/scala/org/clulab/TestUtils.scala
@@ -1,9 +1,10 @@
 package org.clulab
 
 import java.io.File
+
 import org.clulab.processors.Document
 import org.json4s.native.JsonMethods._
-import org.clulab.serialization.json._
+import org.clulab.serialization.json.JSONSerializer
 
 
 object TestUtils {
