+++ odin/src/main/resources/org/clulab/odin/domains/bigmechanism/example/sample.txt
