+++ odin/src/main/scala/org/clulab/odin/impl/DependencyPattern.scala
