+++ odin/src/main/scala/org/clulab/odin/impl/Extractor.scala
