+++ odin/src/main/scala/org/clulab/odin/impl/OdinException.scala
