+++ odin/src/main/scala/org/clulab/odin/impl/OdinResource.scala
