+++ odin/src/main/scala/org/clulab/odin/impl/OdinResourceManager.scala
