+++ odin/src/main/scala/org/clulab/odin/impl/ThompsonVM.scala
