+++ odin/src/main/scala/org/clulab/odin/impl/TokenPattern.scala
