+++ odin/src/main/scala/org/clulab/odin/impl/TokenPatternCompiler.scala
