+++ odin/src/test/resources/org/clulab/odin/embeddings/w2v-gigaword-sample.txt
