+++ odin/src/test/resources/org/clulab/odin/grammars/embeddings.yml
