+++ odin/src/test/resources/org/clulab/odin/variables/master.yml
