+++ odin/src/test/resources/org/clulab/odin/variables/template.yml
