+++ odin/src/test/scala/org/clulab/odin/TestCrossSentencePatterns.scala
