+++ odin/src/test/scala/org/clulab/odin/TestDependencyPattern.scala
