+++ odin/src/test/scala/org/clulab/odin/TestNumericPatterns.scala
