+++ odin/src/test/scala/org/clulab/odin/TestTokenPattern.scala
