+++ odin/src/test/scala/org/clulab/odin/TestUtils.scala
@@ -3,7 +3,7 @@ package org.clulab.odin
 import java.io.File
 import org.clulab.processors.Document
 import org.json4s.native.JsonMethods._
-import org.clulab.serialization.json._
+import org.clulab.odin.serialization.json._
 
 
 object TestUtils {
