+++ odin/src/test/scala/org/clulab/odin/TestVariables.scala
